170: Remove unused files, create /api/about routes for fetching stats
