+++ backend/src/api/about/mod.rs
@@ -1,0 +1,2 @@
+pub mod subreddit;
+pub mod user;
