+++ backend/src/api/mod.rs
@@ -2,6 +2,7 @@ use crate::AppState;
 use axum::{routing::get, Router};
 use std::collections::HashMap;
 
+pub mod about;
 pub mod auth;
 pub mod debug;
 pub mod report;
@@ -13,19 +14,10 @@ type AnyParams = HashMap<String, String>;
 /// Defines routes for the /api path.
 pub fn router() -> Router<AppState> {
     Router::<AppState>::new()
-        .route("/debug/subreddit-about", get(debug::subreddit_about))
+        .route("/about/subreddit", get(about::subreddit::subreddit_about))
+        .route("/about/user", get(about::user::user_about))
+        //
         .route("/debug/post-comments", get(debug::post_comments))
-        .route("/debug/user-about", get(debug::user_about))
         .route("/debug/subreddit-posts", get(debug::subreddit_posts))
         .route("/debug/user-posts", get(debug::user_posts))
-        //
-        .route("/report/sentiment", get(report::sentiment))
-        .route("/report/language", get(report::language))
-        .route("/report/sarcasm", get(report::sarcasm))
-        .route("/report/keywords", get(report::keywords))
-        .route("/report/spam", get(report::spam))
-        .route("/report/politics", get(report::politics))
-        .route("/report/hate-speech", get(report::hate_speech))
-        .route("/report/clickbait", get(report::clickbait))
-        .route("/report/troll", get(report::troll))
 }
