--- backend/src/api/report/clickbait.rs
@@ -1,5 +1,0 @@
-use http::StatusCode;
-
-pub async fn clickbait() -> StatusCode {
-    StatusCode::IM_A_TEAPOT
-}
