--- backend/src/api/report/hate_speech.rs
@@ -1,5 +1,0 @@
-use http::StatusCode;
-
-pub async fn hate_speech() -> StatusCode {
-    StatusCode::IM_A_TEAPOT
-}
