--- backend/src/api/report/keywords.rs
@@ -1,5 +1,0 @@
-use http::StatusCode;
-
-pub async fn keywords() -> StatusCode {
-    StatusCode::IM_A_TEAPOT
-}
