--- backend/src/api/report/language.rs
@@ -1,5 +1,0 @@
-use http::StatusCode;
-
-pub async fn language() -> StatusCode {
-    StatusCode::IM_A_TEAPOT
-}
