--- backend/src/api/report/politics.rs
@@ -1,5 +1,0 @@
-use http::StatusCode;
-
-pub async fn politics() -> StatusCode {
-    StatusCode::IM_A_TEAPOT
-}
