--- backend/src/api/report/sarcasm.rs
@@ -1,5 +1,0 @@
-use http::StatusCode;
-
-pub async fn sarcasm() -> StatusCode {
-    StatusCode::IM_A_TEAPOT
-}
