--- backend/src/api/report/sentiment.rs
@@ -1,5 +1,0 @@
-use http::StatusCode;
-
-pub async fn sentiment() -> StatusCode {
-    StatusCode::IM_A_TEAPOT
-}
