--- backend/src/api/report/spam.rs
@@ -1,5 +1,0 @@
-use http::StatusCode;
-
-pub async fn spam() -> StatusCode {
-    StatusCode::IM_A_TEAPOT
-}
