--- backend/src/api/report/troll.rs
@@ -1,5 +1,0 @@
-use http::StatusCode;
-
-pub async fn troll() -> StatusCode {
-    StatusCode::IM_A_TEAPOT
-}
